refactor: move session handler into session service
