+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/comms/delivery/EmbraceDeliveryService.kt
@@ -144,26 +144,10 @@ internal class EmbraceDeliveryService(
     }
 
     override fun sendCachedSessions(
-        isNdkEnabled: Boolean,
-        ndkService: NdkService,
+        ndkService: NdkService?,
         currentSession: String?
     ) {
         sendCachedCrash()
-        val service = when {
-            isNdkEnabled -> ndkService
-            else -> null
-        }
-        sendCachedSessions(service, currentSession)
-    }
-
-    private fun sendCachedCrash() {
-        val crash = cacheManager.loadCrash()
-        crash?.let {
-            apiService.sendCrash(it)
-        }
-    }
-
-    private fun sendCachedSessions(ndkService: NdkService?, currentSession: String?) {
         backgroundWorker.submit(TaskPriority.HIGH) {
             val allSessions = cacheManager.getAllCachedSessionIds()
 
@@ -174,6 +158,13 @@ internal class EmbraceDeliveryService(
                 }
             }
             sendCachedSessions(allSessions, currentSession)
+        }
+    }
+
+    private fun sendCachedCrash() {
+        val crash = cacheManager.loadCrash()
+        crash?.let {
+            apiService.sendCrash(it)
         }
     }
