+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/session/SessionService.kt
@@ -21,4 +21,6 @@ internal interface SessionService {
      * Ends a session manually. If [clearUserInfo] is true, the user info will be cleared.
      */
     fun endSessionWithManual(clearUserInfo: Boolean)
+
+    fun setSdkStartupInfo(startTimeMs: Long, endTimeMs: Long)
 }
