+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/FakeSessionService.kt
@@ -25,4 +25,8 @@ internal class FakeSessionService : SessionService {
     override fun endSessionWithManual(clearUserInfo: Boolean) {
         manualEndCount++
     }
+
+    override fun setSdkStartupInfo(startTimeMs: Long, endTimeMs: Long) {
+        TODO("Not yet implemented")
+    }
 }
