+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeSessionModule.kt
@@ -5,7 +5,6 @@ import io.embrace.android.embracesdk.FakeSessionService
 import io.embrace.android.embracesdk.injection.SessionModule
 import io.embrace.android.embracesdk.session.BackgroundActivityService
 import io.embrace.android.embracesdk.session.PayloadMessageCollator
-import io.embrace.android.embracesdk.session.SessionHandler
 import io.embrace.android.embracesdk.session.SessionService
 import io.embrace.android.embracesdk.session.orchestrator.SessionOrchestratorImpl
 import io.embrace.android.embracesdk.session.properties.SessionPropertiesService
@@ -15,8 +14,6 @@ internal class FakeSessionModule(
     override val sessionService: SessionService = FakeSessionService(),
     override val sessionPropertiesService: SessionPropertiesService = FakeSessionPropertiesService()
 ) : SessionModule {
-    override val sessionHandler: SessionHandler
-        get() = TODO("Not yet implemented")
 
     override val payloadMessageCollator: PayloadMessageCollator
         get() = TODO("Not yet implemented")
