+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/session/EmbraceSessionServiceTest.kt
@@ -1,25 +1,31 @@
 package io.embrace.android.embracesdk.session
 
+import io.embrace.android.embracesdk.FakeBreadcrumbService
 import io.embrace.android.embracesdk.FakeDeliveryService
 import io.embrace.android.embracesdk.FakeNdkService
+import io.embrace.android.embracesdk.concurrency.BlockingScheduledExecutorService
+import io.embrace.android.embracesdk.fakes.FakeAndroidMetadataService
 import io.embrace.android.embracesdk.fakes.FakeClock
 import io.embrace.android.embracesdk.fakes.FakeConfigService
+import io.embrace.android.embracesdk.fakes.FakeInternalErrorService
+import io.embrace.android.embracesdk.fakes.FakeMemoryCleanerService
+import io.embrace.android.embracesdk.fakes.FakeNetworkConnectivityService
 import io.embrace.android.embracesdk.fakes.FakeProcessStateService
 import io.embrace.android.embracesdk.fakes.FakeTelemetryService
+import io.embrace.android.embracesdk.fakes.FakeUserService
 import io.embrace.android.embracesdk.internal.OpenTelemetryClock
 import io.embrace.android.embracesdk.internal.spans.EmbraceSpansService
+import io.embrace.android.embracesdk.internal.spans.SpansService
+import io.embrace.android.embracesdk.logging.InternalEmbraceLogger
 import io.embrace.android.embracesdk.ndk.NdkService
-import io.embrace.android.embracesdk.payload.Session.LifeEventType
+import io.embrace.android.embracesdk.worker.ScheduledWorker
 import io.mockk.clearAllMocks
-import io.mockk.clearMocks
 import io.mockk.mockk
 import io.mockk.mockkStatic
 import io.mockk.unmockkAll
-import io.mockk.verify
 import org.junit.After
 import org.junit.AfterClass
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Before
 import org.junit.BeforeClass
@@ -31,12 +37,12 @@ internal class EmbraceSessionServiceTest {
     private lateinit var service: EmbraceSessionService
     private lateinit var deliveryService: FakeDeliveryService
     private lateinit var spansService: EmbraceSpansService
+    private lateinit var configService: FakeConfigService
 
     companion object {
 
         private val processStateService = FakeProcessStateService()
         private val ndkService: NdkService = FakeNdkService()
-        private val mockSessionHandler: SessionHandler = mockk(relaxed = true)
         private val clock = FakeClock()
 
         @BeforeClass
@@ -55,6 +61,7 @@ internal class EmbraceSessionServiceTest {
     @Before
     fun before() {
         deliveryService = FakeDeliveryService()
+        configService = FakeConfigService()
         spansService = EmbraceSpansService(
             clock = OpenTelemetryClock(embraceClock = clock),
             telemetryService = FakeTelemetryService()
@@ -68,21 +75,8 @@ internal class EmbraceSessionServiceTest {
 
     @Test
     fun `initializing service should detect early sessions`() {
-        initializeSessionService(ndkEnabled = true, isActivityInBackground = false)
-        assertNotNull(deliveryService.lastSentCachedSession)
-    }
-
-    @Test
-    fun `handle crash successfully`() {
-        initializeSessionService()
-        val crashId = "crash-id"
-
-        // let's start session first so we have an active session
-        service.startSessionWithState(true, clock.now())
-
-        service.endSessionWithCrash(crashId)
-
-        verify { mockSessionHandler.onCrash(crashId) }
+        initializeSessionService(ndkService = ndkService, isActivityInBackground = false)
+        assertNull(deliveryService.lastSentCachedSession)
     }
 
     @Test
@@ -91,79 +85,7 @@ internal class EmbraceSessionServiceTest {
         val coldStart = true
 
         service.startSessionWithState(coldStart, 456)
-        assertEquals("", deliveryService.lastSentCachedSession)
-
-        // verify that a STATE session is started
-        verify {
-            mockSessionHandler.onSessionStarted(
-                coldStart,
-                LifeEventType.STATE,
-                456
-            )
-        }
-    }
-
-    @Test
-    fun `on foreground starts state session successfully for non cold start`() {
-        initializeSessionService()
-        val coldStart = false
-
-        service.startSessionWithState(coldStart, 456)
-
-        // verify that a STATE session is started
-        verify {
-            mockSessionHandler.onSessionStarted(
-                coldStart,
-                LifeEventType.STATE,
-                456
-            )
-        }
-    }
-
-    @Test
-    fun `trigger stateless end session successfully for activity in background`() {
-        initializeSessionService(isActivityInBackground = false)
-        // let's start session first so we have an active session
-        startDefaultSession()
-        clearMocks(mockSessionHandler)
-        service.endSessionWithManual(true)
-
-        // verify session is ended
-        verify(exactly = 1) {
-            mockSessionHandler.onSessionEnded(
-                LifeEventType.MANUAL,
-                0,
-                true
-            )
-        }
-        verify(exactly = 1) {
-            mockSessionHandler.onSessionStarted(
-                false,
-                LifeEventType.MANUAL,
-                0
-            )
-        }
-    }
-
-    @Test
-    fun `trigger stateless end session successfully for activity in foreground`() {
-        initializeSessionService(isActivityInBackground = false)
-        // let's start session first so we have an active session
-        startDefaultSession()
-        val endType = LifeEventType.MANUAL
-
-        service.endSessionWithManual(false)
-
-        // verify session is ended
-        verify { mockSessionHandler.onSessionEnded(endType, 0, false) }
-        // verify that a MANUAL session is started
-        verify {
-            mockSessionHandler.onSessionStarted(
-                false,
-                endType,
-                any()
-            )
-        }
+        assertNull(deliveryService.lastSentCachedSession)
     }
 
     @Test
@@ -172,23 +94,62 @@ internal class EmbraceSessionServiceTest {
         assertNull(spansService.completedSpans())
     }
 
+    @Test
+    fun `simulate session capture enabled after onForeground`() {
+        initializeSessionService()
+
+        // missing start call simulates service being enabled halfway through.
+        service.endSessionWithState(clock.now())
+
+        // nothing is delivered
+        assertEquals(0, deliveryService.lastSentSessions.size)
+
+        // next session is recorded correctly
+        service.startSessionWithState(false, clock.now())
+        clock.tick(10000L)
+        service.endSessionWithState(clock.now())
+        assertEquals(1, deliveryService.lastSentSessions.size)
+    }
+
+    @Test
+    fun `session capture disabled after onForeground`() {
+        initializeSessionService()
+
+        service.startSessionWithState(true, clock.now())
+        clock.tick(10000)
+        // missing end call simulates service being disabled halfway through.
+
+        // nothing is delivered
+        assertEquals(0, deliveryService.lastSentSessions.size)
+
+        service.startSessionWithState(false, clock.now())
+        clock.tick(10000L)
+        service.endSessionWithState(clock.now())
+        assertEquals(1, deliveryService.lastSentSessions.size)
+    }
+
     private fun initializeSessionService(
-        ndkEnabled: Boolean = false,
+        ndkService: NdkService? = null,
         isActivityInBackground: Boolean = true
     ) {
         processStateService.isInBackground = isActivityInBackground
 
         service = EmbraceSessionService(
+            InternalEmbraceLogger(),
+            configService,
+            FakeUserService(),
+            FakeNetworkConnectivityService(),
+            FakeAndroidMetadataService(),
+            FakeBreadcrumbService(),
             ndkService,
-            mockSessionHandler,
+            FakeInternalErrorService(),
+            FakeMemoryCleanerService(),
             deliveryService,
-            ndkEnabled,
-            clock,
-            FakeConfigService()
+            mockk(relaxed = true),
+            mockk(relaxed = true),
+            FakeClock(),
+            SpansService.featureDisabledSpansService,
+            ScheduledWorker(BlockingScheduledExecutorService())
         )
-    }
-
-    private fun startDefaultSession() {
-        service.startSessionWithState(true, clock.now())
     }
 }
